[BREAKING] extract code for input-file to Question obj to question_parser.py and related test
