+++ converter/pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 minversion = 7.0
-; addopts = --cov=src
+addopts = --cov --cov-report=html:converter_coverage
 pythonpath = . src/converter
 testpaths = 
  tests
