+++ converter/src/converter/convert.py
@@ -1,7 +1,9 @@
 import json
 import sys, os
-from translations import translations
+
 from shutil import copy2, rmtree
+
+from question_parser import getAllQuestionsFromFiles
 
 import shutil
 def copy_tree(src, dst, symlinks=False, ignore=None):
@@ -20,112 +22,11 @@ class MyEncoder(json.JSONEncoder):
 	def default(self, o):
 		return o.__dict__
 
-class Question:
-	def __init__(self):
-		self.longQuestion = ""
-		self.shortQuestion = ""
-		self.answer = ""
-		self.category = ""
-		self.sound = ""
-		
-	def setLongQuestion(self, txt):
-		self.longQuestion += txt.strip()
-		self.lastUsed = self.setLongQuestion
-
-	def setShortQuestion(self, txt):
-		self.shortQuestion += txt.strip()
-		self.lastUsed = self.setShortQuestion
-		
-	def setAnswer(self, txt):
-		self.answer += txt.strip()
-		self.lastUsed = self.setAnswer
-		
-	def setImg(self, txt):
-		self.img = txt.strip()
-		
-	def setCategory(self, txt):
-		self.category = txt.strip()
-		
-	def setRound(self, txt):
-		self.round = txt.strip()
-
-	def setSound(self, filename):
-		self.sound = filename.strip()
-
-	def setSourceFile(self, txt):
-		self.sourceFile = txt
-
-	def appendToLastUsedField(self, txt):
-		self.lastUsed(r"\n" + txt.strip())
-
-	def __repr__(self) -> str:
-		return self.round + " - " + self.category + " - " + self.shortQuestion 
-
-def detectLanguage(line):
-	if line.lower().startswith("lange vraag"):
-		return "NL"
-	elif line.lower().startswith("long question"):
-		return "EN"
-	else:
-		return None
-
-#parse the user friendly syntax to Question objects
-def readSingleFile(filename):
-	fileQuestions = []
-	detectedLanguage = None
-
-	
-	print("Reading " + filename)
-	currentQuestion = None
-	with open(filename ,'r', encoding='utf-8') as questionFile:		
-		for line in questionFile:
-			if not ':' in line:
-				if currentQuestion and line != "\n":
-					currentQuestion.appendToLastUsedField(line)
-				continue
-
-			key, value = line.split(':',1)
-			if value.strip() == "/" or value.strip() == "\\":
-				print("Skipping line " + line + " because content is single forward or backward slash")
-				continue
-
-			#lange vraag, or its translation, should be the first line in a block, if you encounter it, finish the current question
-			if not detectedLanguage:
-				detectedLanguage = detectLanguage(line)
-				print(f"detected {detectedLanguage} as the language")
-			if key.lower().strip() == translations[detectedLanguage]["longQuestion"].lower().strip():
-				if(currentQuestion):
-					currentQuestion.setSourceFile(filename)
-					fileQuestions.append(currentQuestion)				
-				currentQuestion = Question()
-				currentQuestion.setLongQuestion(value)
-			elif key.lower().strip() == translations[detectedLanguage]["shortQuestion"].lower().strip():
-				currentQuestion.setShortQuestion(value)
-			elif key.lower().strip() == translations[detectedLanguage]["answer"].lower().strip():
-				currentQuestion.setAnswer(value)
-			elif key.lower().strip() == translations[detectedLanguage]["image"].lower().strip():
-				currentQuestion.setImg(value)
-			elif key.lower().strip() == translations[detectedLanguage]["category"].lower().strip():
-				currentQuestion.setCategory(value)
-			elif key.lower().strip() == translations[detectedLanguage]["round"].lower().strip():
-				currentQuestion.setRound(value)
-			elif key.lower().strip() == translations[detectedLanguage]["sound"].lower().strip():
-				currentQuestion.setSound(value)
-			else:
-				print(f"Skipping line {line}")
-
-	# no more "lange vraag" line means you have to write the current ongoing question
-	if currentQuestion:
-		currentQuestion.setSourceFile(filename)
-		fileQuestions.append(currentQuestion)
-
-	return fileQuestions
-
 def getAllQuestionsOrderedByRoundSpec(unhandledQuestions):
 	import re
 	import copy
-	#loop over the rounds, find if there is a round that has a CATEGORY: spec, and create a new round containing these questions
-	print(unhandledQuestions)
+
+	# Sort questions based on the round they appear in, using the rounds list as a reference
 	unhandledQuestions.sort(key=lambda question: [round.lower() for round in rounds].index(question.round.lower()))
 	questionsToAppend = []
 	for round in rounds:
@@ -147,11 +48,6 @@ def getAllQuestionsOrderedByRoundSpec(unhandledQuestions):
 	unhandledQuestions.sort(key=lambda question: [round.lower() for round in rounds].index(question.round.lower()))
 	return unhandledQuestions
 
-def getAllQuestionsFromFiles(inFiles):
-	allQuestions = []
-	for fileName in inFiles:
-		allQuestions += readSingleFile(fileName)
-	return allQuestions
 
 usageText = """
 USAGE:
@@ -210,7 +106,15 @@ def enablePresenter(title):
 		
 	with open("kwismaster.html", "w") as w:
 		w.write(kwismaster)
-	
+
+
+def detectLanguage(line):
+	if line.lower().startswith("lange vraag"):
+		return "NL"
+	elif line.lower().startswith("long question"):
+		return "EN"
+	else:
+		return None
 
 
 if __name__ == '__main__':
@@ -251,7 +155,12 @@ if __name__ == '__main__':
 	if "specs" in config:
 		specifications = config["specs"]
 
-	allQuestions = getAllQuestionsFromFiles([projectDir + "/" + infile for infile in infiles])
+	allInputFiles = [projectDir + "/" + infile for infile in infiles]
+	with open(allInputFiles[0]) as f:
+		first_line = f.readline().strip('\n')
+	language = detectLanguage(first_line)
+
+	allQuestions = getAllQuestionsFromFiles(allInputFiles, language)
 	questionSortedByRound = getAllQuestionsOrderedByRoundSpec(allQuestions)
 
 	questions = []
